+++ src/interface/pwm.h
@@ -37,9 +37,9 @@ struct Pwm
     };
 
     virtual bool open() = 0;
-    virtual bool set(uint_fast8_t val, const string& sourceName) = 0;
-    virtual void setMin(uint_fast8_t val) = 0;
-    virtual void setMax(uint_fast8_t val) = 0;
+    virtual bool set(int val, const string& sourceName) = 0;
+    virtual void setMin(int val) = 0;
+    virtual void setMax(int val) = 0;
     virtual void setMode(Mode mode) = 0;
     virtual void reset() = 0;
     virtual bool exists() = 0;
